requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
Werkzeug==3.0.1